+++ android/shared/src/commonMain/kotlin/com/yeobaek/data/dto/Book.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Book(
     val authors: List<String>,
-    val bookId: Int,
+    val bookId: Long,
     val passageCount: Int,
     val title: String,
 )
