+++ android/shared/src/commonMain/kotlin/com/yeobaek/data/model/GroupModel.kt
@@ -4,7 +4,7 @@ import com.yeobaek.data.dto.Book
 
 data class GroupModel(
     val book: Book,
-    val clubId: Int,
+    val clubId: Long,
     val memberCount: Int,
     val myProgress: MyProgress?,
     val name: String,
